requirements.txt
altair==5.2.0
attrs==23.1.0
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.4
importlib-metadata==6.11.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
peewee==3.16.3
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.40
protobuf==4.25.1
psutil==5.9.6
psycopg2-binary==2.9.9
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
questionary==1.10.0
readchar==4.0.5
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.7
validators==0.22.0
wcwidth==0.2.8
XlsxWriter==3.1.9
yaspin==2.5.0
zipp==3.17.0
